Use 'long long' in gcd

+++ tool.cpp
@@ -1,8 +1,8 @@
 // Miscellaneous tools
 
 // greatest common divisor
-long gcd(long a, long b) {
-  long r = 1;
+long long gcd(long long a, long long b) {
+  long long r = 1;
   while (r != 0) {
     r = a%b;
     a = b;
